pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "valuation-agent-api"
version = "0.1.0"
description = "Deterministic valuation API for Valuation Agent Workspace"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "quantlib-python>=1.32",
    "python-dateutil>=2.8.0",
    "xlsxwriter>=3.1.0",
    "python-docx>=0.8.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "openapi-python-client>=0.12.0",
]
